+++ GoCqHttp.Sdk/Post/Model/Base/CqMessagePostMovel.cs
@@ -8,7 +8,7 @@ namespace EleCho.GoCqHttpSdk.Post.Model
         public abstract string message_type { get; }
 
         public string sub_type { get; set; }
-        public int message_id { get; set; }
+        public long message_id { get; set; }
         public long user_id { get; set; }
         public CqMsgModel[] message { get; set; }
         public string raw_message { get; set; }
